Revert "workqueue: add is_workqueue_empty() api"
This reverts commit 4e507d05b0843663f8dd421dd429df4bc19a5ef6.

Since the workqueue code deletes the work before executing it,
checking for no work item being currently queued to the workqueue
is not sufficient to guarantee that all the works have finished
execution. Hence, we have moved to using a counter based solution
and so this change is no longer required.

Change-Id: Iad52f4be26917eaeee19fea2706b7f6bf97035cf
Signed-off-by: Pratik Patel <pratikp@codeaurora.org>
Signed-off-by: Raj Kushwaha <rajk@codeaurora.org>